clean up global variables
